pyproject.toml
[project]
name = "pathologydino"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.10.12"
dependencies = [
    "einops>=0.8.1",
    "datasets>=2.20.0",
    "fvcore>=0.1.5.post20221221",
    "iopath>=0.1.10",
    "matplotlib",
    "omegaconf>=2.3.0",
    "opencv-python>=4.12.0.88",
    "pyarrow>=20.0.0",
    "openslide-bin>=4.0.0.8",
    "openslide-python>=1.4.2",
    "scikit-image>=0.25.2",
    "submitit>=1.5.3",
    "torchmetrics>=1.8.0",
    "wandb>=0.21.0",
    "huggingface-hub>=1.0.1",
    "torch==2.8.0",
    "torchvision",
    "xformers>=0.0.32.post2",
    # for Eva benchmarking:
    "imagesize>=1.4.1",
    "scipy>=1.14.0",
    "monai>=1.3.2",
    "lightning>=2.2.0",
    "jsonargparse[omegaconf]>=4.30.0",
    "tensorboard>=2.16.2",
    "loguru>=0.7.2",
    "transformers>=4.38.2",
    "onnxruntime>=1.15.1",
    "onnx>=1.16.0",
    "toolz>=0.12.1",
    "rich>=13.7.1",
    "nibabel>=3.2.2",
    "timm>=1.0.9",
    # for HEST benchmarking:
    "hestcore==1.0.4",
]

[tool.setuptools.packages.find]
include = ["dinov2*"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
override-dependencies = [
    "huggingface-hub>=1.0.1",
]

[tool.uv.pip]
# Use uv's auto detection when installing via `uv pip`
torch-backend = "auto"
